Add required properties to fix oneOf validation
The quantity sample schema was failing to validate when one of the quantity_value types was present because it actually validated for both unit_value and count. The respective properties were not required, meaning that any schema would validate against them. Adding those properties as required now enforces the oneOf constraint properly.